requirements.txt
flask==3.0.2
python-chess==1.999
stockfish==3.28.0
flask-socketio==5.3.6
python-dotenv==1.0.1 